requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
amqp==5.1.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
astroid==2.12.10
    # via pylint
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
billiard==3.6.4.0
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.2.7
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r requirements/base.txt
    #   celery
coverage==6.4.4
    # via -r requirements/test.in
dill==0.3.5.1
    # via pylint
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   djangorestframework
    #   openedx-events
django-crum==0.7.9
    # via -r requirements/base.txt
djangorestframework==3.12.4
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
fastavro==1.6.1
    # via
    #   -r requirements/base.txt
    #   openedx-events
idna==3.4
    # via requests
isort==5.10.1
    # via pylint
kombu==5.2.4
    # via
    #   -r requirements/base.txt
    #   celery
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.7.0
    # via pylint
mock==4.0.3
    # via -r requirements/test.in
openedx-events==1.0.0
    # via -r requirements/base.txt
pbr==5.10.0
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==2.5.2
    # via pylint
prompt-toolkit==3.0.31
    # via
    #   -r requirements/base.txt
    #   click-repl
pycodestyle==2.9.1
    # via -r requirements/test.in
pylint==2.15.3
    # via -r requirements/test.in
pymongo==3.12.3
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytz==2022.2.1
    # via
    #   -r requirements/base.txt
    #   celery
    #   django
requests==2.28.1
    # via -r requirements/test.in
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   click-repl
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.0.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
testfixtures==7.0.0
    # via -r requirements/test.in
tomli==2.0.1
    # via pylint
tomlkit==0.11.5
    # via pylint
typing-extensions==4.3.0
    # via
    #   astroid
    #   pylint
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
wrapt==1.14.1
    # via astroid
